chore: Add save settings functionality to configuration menu
The `saveSettingsButton` element has been added to the configuration menu in `chat.js`. This button triggers the `saveConfig` function, which sends a POST request to the server to save the configuration settings. This commit adds the event listener for the `click` event on the `saveSettingsButton` and calls the `saveConfig` function when clicked.